requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1
django-ckeditor==6.0.0
django-constance==2.7.0
django-debug-toolbar==2.2
django-extensions==3.0.4
django-heroku==0.3.1
django-js-asset==1.2.2
django-picklefield==3.0.1
gunicorn==20.1.0
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
